@@ -8,12 +8,10 @@ import (
 	"bufio"
 	"fmt"
 	"io/fs"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strings"
 
-	"github.com/pkg/errors"
 	"gopkg.in/yaml.v2"
 )
 
@@ -51,7 +49,7 @@ type githubOwners struct {
 func readGithubOwners(codeownersPath string) (*githubOwners, error) {
 	f, err := os.Open(codeownersPath)
 	if err != nil {
-		return nil, errors.Wrapf(err, "failed to open %q", codeownersPath)
+		return nil, fmt.Errorf("failed to open %q: %w", codeownersPath, err)
 	}
 	defer f.Close()
 
@@ -82,7 +80,7 @@ func readGithubOwners(codeownersPath string) (*githubOwners, error) {
 		codeowners.owners[path] = owners
 	}
 	if err := scanner.Err(); err != nil {
-		return nil, errors.Wrapf(err, "scanner error")
+		return nil, fmt.Errorf("scanner error: %w", err)
 	}
 
 	return &codeowners, nil
@@ -128,7 +126,7 @@ func (codeowners *githubOwners) checkManifest(path string) error {
 		return fmt.Errorf("there is no owner for %q in %q", pkgDir, codeowners.path)
 	}
 
-	content, err := ioutil.ReadFile(path)
+	content, err := os.ReadFile(path)
 	if err != nil {
 		return err
 	}
